% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peanutsAddin.R
\name{peanutsAddin}
\alias{peanutsAddin}
\title{Pour packing peanuts from Shiny gadget addin}
\usage{
peanutsAddin()
}
\description{
Launch a Shiny gadget to use \code{\link{pour}} in GUI form. See \code{pour} for details.
}
\examples{
\dontrun{peanutsAddin()}
}
