% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{pourApp}
\alias{pourApp}
\title{Pour packing peanuts from Shiny app}
\usage{
pourApp()
}
\value{
nothing is returned but a Shiny app is launched in the browser
}
\description{
Launch a Shiny app to use \code{\link{pour}} in GUI form.
}
\examples{
\dontrun{pourApp()}
}
