% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{use_app}
\alias{use_app}
\title{Add an existing Shiny app to a source package in development}
\usage{
use_app(app, path = "inst", description = TRUE, readme = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{app}{character, path to app directory.}

\item{path}{character, path app is copied to; should be under source package \code{inst} directory. See details.}

\item{description}{logical, add a \code{DESCRIPTION} file template for showcase mode. See details.}

\item{readme}{logical, add a \code{Readme.md} file template for showcase mode. See details.}

\item{overwrite}{logical, overwrite files. See details.}
}
\description{
Copy an existing Shiny app to a source package (not an installed package) and create \code{DESCRIPTION} and \code{Readme.md} template files related to Shiny showcase mode.
}
\details{
Since this function is for use in a package development context, it is assumed your working directory is the package root directory, hence the default \code{path = "inst"}.
You can create \code{description} and \code{readme} with \code{use_app} and then edit these files.
Alternatively, you can create them directly and with greater control using \code{\link{use_app_description}} and \code{\link{use_app_readme}}.
Also, if a \code{rsconnect} directory exists in \code{app} it is not retained in the new copy in \code{path}.

When \code{overwrite = TRUE}, overwrite a previously added app at \code{path}.
Also overwrite any pre-existing \code{DESCRIPTION} or \code{Readme.md} files copied from the source \code{app} if \code{description = TRUE} or \code{readme = TRUE}.
In the latter case, existing files inside \code{app} are never overwritten. Rather, new template files are created in \code{path} in place of the copied files.
}
\examples{
# Copy an external app to a 'shiny' subdirectory inside package source 'inst' directory.
\dontrun{use_app(app = "../external_app", path = "inst/shiny")}
}
\seealso{
\code{\link{use_app_description}}, \code{\link{use_app_readme}}
}
