% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{use_app_description}
\alias{use_app_description}
\title{Create an app DESCRIPTION file}
\usage{
use_app_description(base_path = ".", title, author = "Author Name",
  url = "https://insert_author_account_here.github.io/",
  license = "MIT", mode = "Showcase", tags)
}
\arguments{
\item{base_path}{output directory.}

\item{title}{app title.}

\item{author}{author name.}

\item{url}{author url.}

\item{license}{license type.}

\item{mode}{Shiny display mode.}

\item{tags}{optional tags. May be a vector. See details.}
}
\description{
Add a DESCRIPTION template to a Shiny app.
}
\details{
This file is used with Shiny app showcase mode.
Most arguments to \code{use_app_description} have defaults you must override, either when calling this function directly,
or making edits to the output after calling \code{\link{use_app}}.
\code{title} and \code{tags} do not have defaults. If missing, they should be updated directly in the generated DESCRIPTION file.
Tags passed in the form \code{tags = c("tag", "another tag")} will result in \code{tag, another-tag}.
}
\examples{
\dontrun{use_app_description()}
}
\seealso{
\code{\link{use_app}}, \code{\link{use_app_readme}}
}
