% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{use_app_readme}
\alias{use_app_readme}
\title{Create an app Readme.md file}
\usage{
use_app_readme(base_path = ".")
}
\arguments{
\item{base_path}{output directory.}
}
\description{
Add a Readme.md template to a Shiny app.
}
\details{
This file is used to add a text description below a Shiny app when run in showcase mode.
}
\examples{
\dontrun{use_app_readme()}
}
\seealso{
\code{\link{use_app}}, \code{\link{use_app_description}}
}
