% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{use_clone_comments}
\alias{use_clone_comments}
\title{Add \code{clone-comments.md} to package}
\usage{
use_clone_comments(lint_as_test = FALSE)
}
\arguments{
\item{lint_as_test}{logical, if set to \code{TRUE} a statement is added to the file regarding use of \code{lintr} with \code{testthat}.}
}
\value{
side effect of creating file.
}
\description{
Add \code{clone-comments.md} to package if there are special instructions or information other developers cloning the package repository should know.
}
\examples{
\dontrun{use_clone_comments}
}
