% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{use_hex}
\alias{use_hex}
\title{Add \code{hex.R} to package}
\usage{
use_hex(account, host = "github")
}
\arguments{
\item{account}{user account.}

\item{host}{\code{"github"} or \code{"bitbucket"}.}
}
\value{
side effect of creating file.
}
\description{
Add \code{hex.R} to package \code{data-raw} directory. Create a default logo.
}
\details{
This script provides a template for package hex sticker icons, a large icon in \code{data-raw} and a small icon in \code{inst}.
A default logo is generated and placed at \code{man/figures/logo.png}. However, it will require user customization after it is generated.
Adapt the provided script and rerun to make a new logo.

The default \code{logo.png} will not be created if the \code{magick} package is not installed.
Instead, a message is printed notifying of this requirement. Being able to generate a default logo (that will surely be replaced later) is a minor, optional feature.
Therefore, \code{pkgpeanuts} does not have package dependencies in this regard.
}
\examples{
\dontrun{use_hex(account = "username", host = "github")}
}
