% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{use_lintr}
\alias{use_lintr}
\title{Add \code{lintr} to R package}
\usage{
use_lintr(base_path = ".", lint_as_test = FALSE)
}
\arguments{
\item{base_path}{package root directory.}

\item{lint_as_test}{logical, if \code{TRUE} (default), adds \code{tests/testthat/test-lintr.R} to package. See details.}
}
\value{
side effect of adding \code{lintr} scaffolding to package.
}
\description{
Add scaffolding for \code{lintr} package usage to R package.
}
\details{
If \code{lint_as_test = FALSE} (default), a \code{.lintr} file is created in the package root directory.
Set to \code{TRUE} if you intend to use \code{lintr} with unit testing so that problems caught by the linter cause test failure.
Do not set to \code{TRUE} simply because you are using \code{testthat}, but apart from \code{lintr}.

When \code{lint_as_test = TRUE}, this function creates the template \code{.lintr} file in the \code{inst} directory and
adds a symbolic link to this file in the package root directory.
It also adds both files (as applicable) to \code{.Rbuildignore}. It also creates a unit test for lint-free package code only when \code{lint_as_test = TRUE}.

NOTE: You will have to run R as Administrator on Windows when \code{lint_as_test = TRUE} or you it the function will fail to create the symbolic link.
Similar issues could present on other systems depending on user permissions.
}
\examples{
\dontrun{use_lintr()}
}
